[protocol 3.6] Uint96 math

+++ packages/loopring_v3/contracts/core/impl/libtransactions/AccountUpdateTransaction.sol
@@ -32,7 +32,7 @@ library AccountUpdateTransaction
         address owner;
         uint32  accountID;
         uint16  feeTokenID;
-        uint    fee;
+        uint96  fee;
         uint    publicKey;
         uint32  validUntil;
         uint32  nonce;
